update
- tested version

+++ STM32/NUCLEO-L432KC/L2_Timer_Switch-Case/Core/Src/my_controls.c
@@ -9,7 +9,7 @@
 #include "tim.h"
 #include "stdbool.h"
 
-#include "my_controls.h"
+#include <my_controls.h>
 #include <my_signalshapes.h>
 
 
@@ -19,7 +19,7 @@
  * initialization of counter variable
  * ----------------------------------------------------------------------
  */
-uint16_t cnt = 0;
+uint8_t cnt = 0;
 
 /*
  * ----------------------------------------------------------------------
